pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strands-agents-portkey"
version = "0.1.0"
description = "Portkey model provider for Strands Agents"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Federico Kamelhar"},
    {name = "Portkey AI, Inc."},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "strands-agents>=0.1.0",
    "portkey-ai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "pytest-asyncio>=1.0.0,<1.2.0",
    "ruff>=0.12.0,<0.13.0",
    "mypy>=1.15.0,<2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/strands_portkey"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "pytest-asyncio>=1.0.0,<1.2.0",
    "ruff>=0.12.0,<0.13.0",
    "mypy>=1.15.0,<2.0.0",
]

[tool.hatch.envs.default.scripts]
test       = "pytest tests/unit/ --cov=strands_portkey --cov-report=term-missing -q"
test-integ = "pytest tests/integration/ -v"
lint       = "ruff check src/ tests/unit/"
fmt        = "ruff format src/ tests/"
type-check = "mypy src/"
check      = ["lint", "type-check", "test"]

[tool.ruff]
line-length = 120
include = ["src/**/*.py", "tests/**/*.py", "examples/**/*.py"]

[tool.ruff.lint]
select = [
  "B",   # flake8-bugbear
  "D",   # pydocstyle
  "E",   # pycodestyle
  "F",   # pyflakes
  "I",   # isort
]

[tool.ruff.lint.per-file-ignores]
"!src/**/*.py" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
